chore: narrowing scope
