+++ src/libecalc/presentation/yaml/mappers/process/mapped_section_validator.py
@@ -7,8 +7,6 @@ from libecalc.process.process_units.compressor import Compressor
 from libecalc.process.process_units.mixer import Mixer
 from libecalc.process.process_units.splitter import Splitter
 
-_CHOKE = ("UPSTREAM_CHOKE", "DOWNSTREAM_CHOKE")
-
 _PRESSURE_CONTROL_REQUIRES_ANTI_SURGE = {
     "COMMON_ASV": AntiSurgeType.COMMON_ASV,
     "INDIVIDUAL_ASV_RATE": AntiSurgeType.INDIVIDUAL_ASV,
@@ -21,9 +19,9 @@ class MappedSectionValidator:
 
     def validate(self, sections: Sequence[MappedSection]) -> None:
         for s in sections:
+            self._validate_section_has_compressor(s)
             self._validate_strategy_compatibility(s)
             self._validate_anti_surge_vs_compressors(s)
-            self._validate_pressure_control_vs_compressors(s)
             self._validate_common_asv_units(s)
 
     @staticmethod
@@ -47,13 +45,10 @@ class MappedSectionValidator:
                 f"'{s.constraint.anti_surge}' cannot be used for {location}. Sections without cannot have anti-surge."
             )
 
-    def _validate_pressure_control_vs_compressors(self, s: MappedSection) -> None:
-        location = self._section_location(s)
-        has_compressors = any(isinstance(u, Compressor) for u in s.process_units)
-        if not has_compressors and s.constraint.pressure_control not in _CHOKE:
+    def _validate_section_has_compressor(self, s: MappedSection) -> None:
+        if not any(isinstance(u, Compressor) for u in s.process_units):
             raise EcalcValidationException(
-                f"'{s.constraint.pressure_control}' cannot be used for {location} because the section does not contain a compressor."
-                f"Sections without compressors must use UPSTREAM_CHOKE or DOWNSTREAM_CHOKE."
+                f"Process sections without compressors are not supported yet for {self._section_location(s)}."
             )
 
     def _validate_common_asv_units(self, s: MappedSection) -> None:
